requirements.txt
Babel~=2.9.1
blinker==1.4
boto3~=1.21.0
cffi~=1.15.0
cryptography~=36.0.1
factory-boy~=3.2.1
Faker~=12.3.3
firebase-admin~=5.2.0
Flask~=2.0.3
Flask-Cors~=3.0.10
Flask-Limiter~=2.1.3
Flask-Migrate~=3.1.0
Flask-Seasurf~=0.3.1
Flask-SQLAlchemy~=2.5.1
flask-talisman~=0.8.1
grpcio==1.43.0
grpcio-tools==1.43.0
marshmallow~=3.14.1
psycopg2~=2.9.3
pycparser==2.21
pytest~=7.0.1
python-dotenv~=0.19.2
requests~=2.27.1
rollbar~=0.16.2
SQLAlchemy~=1.4.31
webargs~=8.1.0
